Deploy.sh now only links not existing files/dirs. Warns when exists and is not a link.

+++ deploy.sh
@@ -3,11 +3,17 @@ setopt extended_glob
 
 # All directories not disabled
 for dir in `ls -d *~*.disabled(/)`; do
-  echo $dir
   for target in `ls $dir`; do
-    echo ".. " $target
-    rm -irf $HOME/.$target
-    ln -s $HOME/dotfiles/$dir/$target $HOME/.$target
+    if [ ! -L $HOME/.$target ]; then
+      if [ -f $HOME/.$target ]; then
+        echo "[Warning]:" $target " is a regular file."
+      elif [ -d $HOME/.$target ]; then
+        echo "[Warning]:" $target " is a regular directory."
+      else
+        echo "Linking... " $target
+        ln -s $HOME/dotfiles/$dir/$target $HOME/.$target
+      fi
+    fi
   done
 done
